Finally fix the bug that has been taunting me
It's a very specific bug that only happens when you have a comment, on the same line as active code, in a 1 line block statement (excluding the opening and closing delimiters) which made readLine in tokenizer read my test.jiv input wrong as it lost MultipleLinesOutput because it thought it had to parse inner code. This only happened when you had this comment, on thew same line as code, in a block inside another block. Which is hy i only had it very few times and thought it was just a one time coincidence.

The fix was to before spitting into an array and cheking that array, take that rray and put it into anew helper method which decimates single line comments early. Avoids the hassle when needed.