Add support for importing and exporting symbols
To export a symbol, by its name add a *, then to import symbols in the file you want to use it in type "tsea "filePath"!" to get the edxported symbols or "tsea "filePath" <- funca, funcg!" to get a specific set of symbols.

TODO: Also make the extneison support this update.